+++ CHANGELOG.md
@@ -13,3 +13,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Retrieval of locations, climbs and attempts from firebase
 - Views for locations and climbs
 - Location, sections and grade set creation
+- Location editing